rpm packaging: allow any version of tar
Relax the tar dependency by allowing any version of tar to be used.
Initially, the requirement for tar < v1.27 was because the tar gnu
format changed a bit between v1.26 and v1.27 (more specifically the
mode/time fields of @LongLink headers). Thus, the delta might not apply
and tarballs committed (with pristine-tar) on a system with old tar
could possibly not be checked out on a system with newer tar (and vice
versa). Debian has a patched tar program (PRISTINE_TAR_COMPAT patch)
that restores the old tar format when pristine-tar is used (pristine-tar
sets the PRISTINE_TAR_COMPAT environment variable which the tar program
looks at).

However, it turns out that we do not suffer this problem (and do not
need patched tar) in Tizen tools. We have for a long time had
"Use-posix-tar-format-by-default.patch" which makes pristine-tar to
always use the posix tar format, regardless of the underlying system.
Thus, virtually all pristine-tar delta in Tizen packages should have
been generated using the posix tar format (which has not changed) and
delta should apply cleanly.

Change-Id: Ib0f203f69dc4ccb7a9ca9e8346d69522c0507734
Signed-off-by: Markus Lehtonen <markus.lehtonen@linux.intel.com>